LightVisualiserAlgo : Add drawing methods
These are moved from `StandardLightVisualiser` to make some previously
private methods accessible. It also allows light visualisers to derive
directly from `LightVisualiser` instead of needing to derive from
`StandardLightVisualiser` just to get the static drawing methods.
