+++ include/GafferSceneUI/StandardLightVisualiser.h
@@ -68,16 +68,6 @@ class GAFFERSCENEUI_API StandardLightVisualiser : public IECoreGLPreview::LightV
 
 	protected :
 
-		static IECoreGL::ConstRenderablePtr ray( bool muted = false );
-		static IECoreGL::ConstRenderablePtr pointRays( float radius = 0, bool muted = false );
-		static IECoreGL::ConstRenderablePtr distantRays( bool muted = false );
-		static IECoreGL::ConstRenderablePtr spotlightCone( float innerAngle, float outerAngle, float lensRadius, float length = 1.0f, float lineWidthScale = 1.0f, bool muted = false );
-		static IECoreGL::ConstRenderablePtr sphereWireframe( float radius, const Imath::Vec3<bool> &axisRings, float lineWidthScale = 1.0f, const Imath::V3f &center = Imath::V3f( 0.0f ), bool muted = false );
-
-		static IECoreGL::ConstRenderablePtr quadPortal( const Imath::V2f &size, float hatchingScale = 1.0f, bool muted = false );
-
-		static IECoreGL::ConstRenderablePtr colorIndicator( const Imath::Color3f &color );
-
 		// This method should be overridden by any sub-classes that wish to
 		// provide an alternate surface texture for area-based lights.
 		//
@@ -93,24 +83,6 @@ class GAFFERSCENEUI_API StandardLightVisualiser : public IECoreGLPreview::LightV
 		virtual IECore::DataPtr surfaceTexture( const IECore::InternedString &attributeName, const IECoreScene::ShaderNetwork *shaderNetwork, const IECore::CompoundObject *attributes, int maxTextureResolution ) const;
 
 	private :
-
-		/// \todo Expose publicly once we have enough uses to dictate
-		/// the most general set of parameters.
-		static IECoreGL::ConstRenderablePtr pointShape( float radius, bool muted = false );
-
-		// textureData should be as per return type of surfaceTexture
-
-		static IECoreGL::ConstRenderablePtr environmentSphereSurface( IECore::ConstDataPtr textureData, const Imath::Color3f &tint, int textureMaxResolution, const Imath::Color3f &fallbackColor );
-
-		static IECoreGL::ConstRenderablePtr quadWireframe( const Imath::V2f &size, bool muted = false );
-		static IECoreGL::ConstRenderablePtr quadSurface( const Imath::V2f &size, IECore::ConstDataPtr textureData, const Imath::Color3f &tint, int textureMaxResolution, const Imath::Color3f &fallbackColor, const Imath::M33f &uvOrientation );
-
-		static IECoreGL::ConstRenderablePtr diskWireframe( float radius, bool muted = false );
-		static IECoreGL::ConstRenderablePtr diskSurface( float radius, IECore::ConstDataPtr textureData, const Imath::Color3f &tint, int textureMaxResolution, const Imath::Color3f &fallbackColor );
-
-		static IECoreGL::ConstRenderablePtr cylinderWireframe( float radius, float length, bool muted = false );
-		static IECoreGL::ConstRenderablePtr cylinderSurface( float radius, float length, const Imath::Color3f &color );
-		static IECoreGL::ConstRenderablePtr cylinderRays( float radius, bool muted = false );
 
 		static LightVisualiser::LightVisualiserDescription<StandardLightVisualiser> g_description;
